@@ -16,16 +16,16 @@ import java.util.Map;
 import java.util.TreeMap;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
-import javax.annotation.PreDestroy;
+import jakarta.annotation.PreDestroy;
 import javax.xml.XMLConstants;
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBException;
-import javax.xml.bind.Marshaller;
-import javax.xml.bind.PropertyException;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.JAXBContext;
+import jakarta.xml.bind.JAXBException;
+import jakarta.xml.bind.Marshaller;
+import jakarta.xml.bind.PropertyException;
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
@@ -151,7 +151,12 @@ public class XmlReporterFactory extends Format {
         try {
             marshaller.setProperty("com.sun.xml.internal.bind.xmlHeaders", header);
         } catch (PropertyException pex) {
-            marshaller.setProperty("com.sun.xml.bind.xmlHeaders", header);
+            try {
+                marshaller.setProperty("com.sun.xml.bind.xmlHeaders", header);
+            } catch (PropertyException pex2) {
+                // taken from org.glassfish.jaxb.runtime.v2.runtime.MarshallerImpl.setProperty
+                marshaller.setProperty("org.glassfish.jaxb.xmlHeaders", header);
+            }
         }
     }
 
