@@ -43,7 +43,7 @@ import java.time.ZonedDateTime
 import static org.mockito.Mockito.mock
 import static org.mockito.ArgumentMatchers.any
 
-import javax.annotation.PostConstruct
+import jakarta.annotation.PostConstruct
 import java.util.concurrent.RejectedExecutionException
 
 @Configuration
@@ -130,7 +130,7 @@ class TestConfig {
             .build()
 
         RollingLogResolver rollingLogResolver = new SimpleRollingLogResolver()
-        return new SshTemplate(passwordSetting, env.getProperty("ssh.timestampFormat"), rollingLogResolver)
+        return new SshTemplate(passwordSetting, env.getProperty("ssh.timestampFormat"), rollingLogResolver, false)
     }
 
     @Bean
