@@ -5,7 +5,7 @@ import org.jbehavesupport.core.internal.web.YamlElementLocatorParser
 import org.jbehavesupport.core.internal.web.by.ByFactoryResolverImpl
 import org.jbehavesupport.core.internal.web.by.CssByFactory
 import org.jbehavesupport.core.internal.web.by.XpathByFactory
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 import static org.openqa.selenium.By.cssSelector
 import static org.openqa.selenium.By.xpath
