@@ -1,15 +1,12 @@
 package org.jbehavesupport.core.healthcheck;
 
-import static java.util.stream.Collectors.toList;
-
-import java.util.ArrayList;
 import java.util.List;
 
 import lombok.RequiredArgsConstructor;
+import org.assertj.core.api.SoftAssertions;
 import org.jbehave.core.annotations.Given;
 import org.jbehave.core.model.ExamplesTable;
 import org.jbehave.core.steps.Row;
-import org.junit.runners.model.MultipleFailureException;
 import org.springframework.beans.factory.NoSuchBeanDefinitionException;
 import org.springframework.beans.factory.annotation.BeanFactoryAnnotationUtils;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
@@ -48,26 +45,23 @@ public final class HealthCheckSteps {
     private final ConfigurableListableBeanFactory beanFactory;
 
     @Given("these components are healthy:$componentList")
-    public void checkComponentsAreHealthy(ExamplesTable componentList) throws MultipleFailureException {
+    public void checkComponentsAreHealthy(ExamplesTable componentList) {
         List<String> componentQualifiers = componentList.getRowsAsParameters().stream()
             .map(Row::values)
             .map(row -> row.get("component"))
-            .collect(toList());
+            .toList();
 
-        List<Throwable> exceptions = new ArrayList<>();
+        SoftAssertions softly = new SoftAssertions();
         for (String componentQualifier : componentQualifiers) {
             HealthCheck healthCheck = resolveHealthCheck(componentQualifier);
             try {
                 healthCheck.check();
             } catch (Exception e) {
-                IllegalStateException ie = new IllegalStateException("Component " + componentQualifier + " is not healthy. " + e.getMessage(), e);
-                exceptions.add(ie);
+                softly.fail("Component " + componentQualifier + " is not healthy. " + e.getMessage(), e);
             }
         }
 
-        if (!exceptions.isEmpty()) {
-            throw new MultipleFailureException(exceptions);
-        }
+        softly.assertAll();
     }
 
     private HealthCheck resolveHealthCheck(String qualifier) {
