@@ -1,6 +1,5 @@
 package org.jbehavesupport.core.internal.web.webdriver;
 
-import io.github.bonigarcia.wdm.WebDriverManager;
 import org.jbehavesupport.core.web.WebDriverFactory;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
@@ -10,7 +9,7 @@ import org.openqa.selenium.remote.RemoteWebDriver;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import java.util.concurrent.TimeUnit;
+import java.time.Duration;
 
 @Component
 public class ChromeWebDriverFactory implements WebDriverFactory {
@@ -23,16 +22,17 @@ public class ChromeWebDriverFactory implements WebDriverFactory {
     @Value("${web.browser.driver.location:#{null}}")
     private String browserDriverLocation;
 
+    @Value("${web.browser.version:#{null}}")
+    private String browserVersion;
+
     @Value("${web.browser.driver.port:#{null}}")
     private Integer browserPort;
-
-    @Value("${web.browser.driver.version:#{null}}")
-    private String browserDriverVersion;
 
     @Value("${web.browser.driver.startup.arguments:#{null}}")
     private String browserStartupArguments;
 
-    private boolean driverSetup = false;
+    @Value("${web.browser.binary.location:#{null}}")
+    private String binaryLocation;
 
     private RemoteWebDriver driver = null;
 
@@ -44,31 +44,31 @@ public class ChromeWebDriverFactory implements WebDriverFactory {
     @Override
     public RemoteWebDriver createWebDriver() {
         createChromeDriver();
-        driver.manage().timeouts().implicitlyWait(timeout, TimeUnit.SECONDS);
-        driver.manage().timeouts().pageLoadTimeout(timeout, TimeUnit.SECONDS);
-        driver.manage().timeouts().setScriptTimeout(timeout, TimeUnit.SECONDS);
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(timeout));
+        driver.manage().timeouts().pageLoadTimeout(Duration.ofSeconds(timeout));
+        driver.manage().timeouts().scriptTimeout(Duration.ofSeconds(timeout));
         return driver;
     }
 
     private void createChromeDriver() {
-        if (!driverSetup) {
-            if (browserDriverLocation != null) {
-                System.setProperty("webdriver.chrome.driver", browserDriverLocation);
-            } else {
-                WebDriverManager browserManager = WebDriverManager.chromedriver();
-                if (browserDriverVersion != null) {
-                    browserManager.driverVersion(browserDriverVersion);
-                }
-                browserManager.setup();
-            }
-            driverSetup = true;
+        if (browserDriverLocation != null) {
+            System.setProperty("webdriver.chrome.driver", browserDriverLocation);
         }
 
         ChromeOptions options = new ChromeOptions();
+
+        if (browserVersion != null){
+            options.setBrowserVersion(browserVersion);
+        }
+
         if (browserStartupArguments != null) {
             options.addArguments(browserStartupArguments.split("\\s+"));
         } else {
             options.addArguments("--start-maximized");
+        }
+
+        if (binaryLocation != null) {
+            options.setBinary(binaryLocation);
         }
 
         ChromeDriverService.Builder driverServiceBuilder = new ChromeDriverService.Builder();
