@@ -1,24 +1,20 @@
 package org.jbehavesupport.core.rest
 
+import org.jbehavesupport.test.support.TestAppSupport
 import org.jbehavesupport.test.support.TestSupport
-import org.junit.runner.JUnitCore
-import spock.lang.Shared
 import spock.lang.Specification
 
-class NegativeRestScenariosIT extends Specification implements TestSupport {
-
-    @Shared
-        runner = new JUnitCore()
+class NegativeRestScenariosIT extends Specification implements TestSupport, TestAppSupport {
 
     def "Should validate REST response via success handlers"() {
         when:
-        def result = runner.run(runWith("rest/NegativeRestHandling.story"))
+        def result = run(runWith("rest/NegativeRestHandling.story"))
 
         then:
         true
-        result.getFailureCount() == 3
-        result.getFailures().get(0).getMessage().contains("Expected response code is CREATED but was BAD_REQUEST")
-        result.getFailures().get(1).getMessage().contains("value 'sad' is not equal to 'happy'")
-        result.getFailures().get(2).getMessage().contains("value 'astronaut' is not equal to 'hippopotamus'")
+        result.getTotalFailureCount() == 3
+        result.getFailures().get(0).exception.getMessage().contains("Expected response code is CREATED but was BAD_REQUEST")
+        result.getFailures().get(1).exception.getMessage().contains("value 'sad' is not equal to 'happy'")
+        result.getFailures().get(2).exception.getMessage().contains("value 'astronaut' is not equal to 'hippopotamus'")
     }
 }
