@@ -1,19 +1,14 @@
 package org.jbehavesupport.core.ws
 
+import org.jbehavesupport.test.support.TestAppSupport
 import org.jbehavesupport.test.support.TestSupport
-import org.junit.runner.JUnitCore
-import spock.lang.Issue
-import spock.lang.Shared
 import spock.lang.Specification
 
-class RequestResponseAliasIT extends Specification implements TestSupport {
-
-    @Shared
-        runner = new JUnitCore()
+class RequestResponseAliasIT extends Specification implements TestSupport, TestAppSupport {
 
     def "Test us request response aliases"() {
         when:
-        def result = runner.run(runWith("ws/RequestResponseAlias.story"))
+        def result = run(runWith("ws/RequestResponseAlias.story"))
 
         then:
         result.failures.isEmpty()
