@@ -2,27 +2,25 @@ package org.jbehavesupport.test.sample.report;
 
 import org.jbehavesupport.core.AbstractSpringStories;
 import org.jbehavesupport.test.support.SshContainer;
-import org.jbehavesupport.test.support.TestConfig;
-import org.junit.ClassRule;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.ContextConfiguration;
+import org.jbehavesupport.test.support.TestAppContainer;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.Arrays;
 import java.util.List;
 
-// spring junit4 runner needed for testcontainer lifecycle by classrule to work
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = TestConfig.class)
 public class ReportSampleStoriesIT extends AbstractSpringStories {
 
-    @ClassRule
     public static SshContainer sshContainer = new SshContainer();
+    public static TestAppContainer testAppContainer = TestAppContainer.getTestAppContainer();
+
+    static {
+        sshContainer.start();
+        testAppContainer.start();
+    }
 
     @DynamicPropertySource
-    static void sshProperties(DynamicPropertyRegistry registry) {
+    static void containerProperties(DynamicPropertyRegistry registry) {
         sshContainer.updateDynamicPropertyRegistry(registry);
     }
 
