@@ -4,8 +4,6 @@ import lombok.RequiredArgsConstructor;
 import net.schmizz.sshj.SSHClient;
 import net.schmizz.sshj.transport.verification.PromiscuousVerifier;
 import org.jbehave.core.annotations.Given;
-import org.jbehavesupport.core.ssh.SshTemplate;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Component;
 
@@ -21,27 +19,25 @@ import java.time.format.DateTimeFormatter;
 @RequiredArgsConstructor
 public final class SshGeneratorSteps {
 
-    @Qualifier("TEST")
-    private final SshTemplate sshTemplate;
-
     private final Environment environment;
 
     @Given("ssh test data are filled")
     public void fillTestData() throws IOException {
-        SSHClient sshClient = new SSHClient();
-        sshClient.addHostKeyVerifier(new PromiscuousVerifier());
-        sshClient.connect(environment.getProperty("ssh.hostname"), environment.getProperty("ssh.port", Integer.class));
-        sshClient.authPassword(environment.getProperty("ssh.credentials.user"), environment.getProperty("ssh.credentials.password"));
+        try (SSHClient sshClient = new SSHClient()) {
+            sshClient.addHostKeyVerifier(new PromiscuousVerifier());
+            sshClient.connect(environment.getProperty("ssh.hostname"), environment.getProperty("ssh.port", Integer.class));
+            sshClient.authPassword(environment.getProperty("ssh.credentials.user"), environment.getProperty("ssh.credentials.password"));
 
-        ZonedDateTime startTime = ZonedDateTime.now();
-        String timestampFormat = environment.getProperty("ssh.timestampFormat");
-        String timestamp = startTime.withZoneSameInstant(ZoneId.of("GMT")).format(DateTimeFormatter.ofPattern(timestampFormat));
-        String logText = "some long string containing cdata in many Cdata forms." +
-            "Such as correct one <![CDATA[1832300759061]]> and malformed <![[1832300759061]]> " +
-            "and incomplete <![CDATA[ and duplicated correct one <![CDATA[1832300759061]]> with some additional information" +
-            "Also some unexpected closing like ] and ]] also sharp ]]>";
-        String command = "echo " + timestamp + " " + "\"" + logText + "\"" + " > " + environment.getProperty("ssh.logPath");
-        sshClient.startSession().exec(command);
+            ZonedDateTime startTime = ZonedDateTime.now();
+            String timestampFormat = environment.getProperty("ssh.timestampFormat");
+            String timestamp = startTime.withZoneSameInstant(ZoneId.of("GMT")).format(DateTimeFormatter.ofPattern(timestampFormat));
+            String logText = "some long string containing cdata in many Cdata forms." +
+                "Such as correct one <![CDATA[1832300759061]]> and malformed <![[1832300759061]]> " +
+                "and incomplete <![CDATA[ and duplicated correct one <![CDATA[1832300759061]]> with some additional information" +
+                "Also some unexpected closing like ] and ]] also sharp ]]>";
+            String command = "echo " + timestamp + " " + "\"" + logText + "\"" + " > " + environment.getProperty("ssh.logPath");
+            sshClient.startSession().exec(command);
+        }
     }
 
 }
